// apps/desktop/src/generated/query/TopologyDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RelationDto } from "./RelationDto";
import type { ResourceDto } from "./ResourceDto";
import type { SnapshotMetadata } from "./SnapshotMetadata";
import type { TopologyFrontierDto } from "./TopologyFrontierDto";

export type TopologyDto = { metadata: SnapshotMetadata, focus_resource_id: string, depth: number, nodes: Array<ResourceDto>, edges: Array<RelationDto>, frontier: Array<TopologyFrontierDto>, truncated: boolean, };